Fix crash if security prevents directory read
Add exception handling to catch IOException and
UnauthorizedAccessException when enumerating directories. These
exceptions will be common enough that they should probably be handled
in the ScanDirectory method rather than rethrown. Adding to the error
or warning log and continuing the scan seems like the most prudent
solution for the time being.