Fix bug with quoted command line path arguments
Path arguments are now checked for a trailing double quote character.
If found, the double quote is removed. If a path argument enclosed in
double quotes ended with a backslash, the last two characters would be
treated as the escape character \". This caused a trailing quote
character to be included on any such path argument, which resulted in
invalid path errors.