Remove IEquatable implementation from ScanError
After writing tests for ScanError and evaluating the current use of the Equals methods and equality/inequality operators throughout the codebase, it was discovered that these methods aren't in use and would have questionable utility in the future. In addition, unit test development revealed that the IEquatable implementation was flawed due to the difficulty in making any kind of memberwise or otherwise semantic comparison between ScanError.Exception properties. For these reasons, the IEquatable implementation has been removed from ScanError in addition to any associated tests.