FM-69: FileScanning unit tests (#39)
* Move FileScannerTests.cs to Orchestration directory

FileScannerTests.cs contains old tests for file scanner classes that were eventually refactored into *Orchestration classes. Many of the tests may be useful for creating tests on the new classes.

* Add ScanErrorTests, cleanup other tests and test list

* Remove null check in ScanError equality operator

Removed the check for left and right operands being null from equality operator. This condition is covered by the ReferenceEquals call.

* Remove IEquatable implementation from ScanError

After writing tests for ScanError and evaluating the current use of the Equals methods and equality/inequality operators throughout the codebase, it was discovered that these methods aren't in use and would have questionable utility in the future. In addition, unit test development revealed that the IEquatable implementation was flawed due to the difficulty in making any kind of memberwise or otherwise semantic comparison between ScanError.Exception properties. For these reasons, the IEquatable implementation has been removed from ScanError in addition to any associated tests.

* Add ScanError unit tests, cleanup code

* Add null checks to ScanResult addition operator

Added null check to ScanResult addition operator that throws an exception if both operands are null. Also modified the method to allow for one of the two operands to be null.

* Add ScanResultTests

---------

Co-authored-by: Jason Sells <jasonsells@protonmail.com>