FM-70 - orchestration unit tests (partial) (#41)
* Cleanup tests

Changed test descriptions and method names to match new style that I hopefully stick with for more than a week.

* Add DirectoryScanOptions unit tests

* Update testlist

* Update argument checking in DirectoryScanOrchestrator, update tests

* Remove unneeded parameter check in DirectoryScanOrchestrator

Null check for DirectoryScanOptions.Directory is now done in the constructor so we can fail faster if no scan directory is provided.

Updated DirectoryScanOrchestrator unit tests.

* Add comment regarding possible issue in DirectoryScanOrchestrator

See comment for details.

* Add empty file list check to DirectoryScanOrchestrator.ScanDirectoryAsync

If file list is empty, we can return early instead of calling the FileScanOrchestrator.

Added unit tests for DirectoryScanOrchestrator.

* Add null parameter check to FileScanOrchestrator.ScanFilesAsync

* Update FileScanOrchestratorTests

* Update FileScanOrchestrator tests

* Separate ScanFilesAsync unit test

Need to validate two calls for ScanFilesAsync (ComputeHashFromStreamAsync and AddAsync). ComputeHashFromStreamAsync is difficult to properly validate because I should validate that it's being called with Streams containing the correct data. Will return to this.